small issues

+++ src/command_line.c
@@ -96,14 +96,14 @@ static bool is_valid_argument(char *arg)
 static void open_console()
 {
 	int hConHandle;
-    long lStdHandle;
+    intptr_t lStdHandle;
     FILE *fp;
 
     // Allocate a console for this app
     AllocConsole();
 
     // Redirect unbuffered STDOUT to the console
-    lStdHandle = (long)GetStdHandle(STD_OUTPUT_HANDLE);
+    lStdHandle = (intptr_t)GetStdHandle(STD_OUTPUT_HANDLE);
     hConHandle = _open_osfhandle(lStdHandle, _O_TEXT);
     fp = _fdopen(hConHandle, "w");
     *stdout = *fp;
